requirements.txt
fastapi
uvicorn
streamlit
requests
google-generativeai
chromadb-client
python-dotenv